Add show file
