+++ app/controllers/admin/application_controller.rb
@@ -8,11 +8,13 @@
 # you're free to overwrite the RESTful controller actions.
 module Admin
   class ApplicationController < Administrate::ApplicationController
-    before_action :authenticate_admin
+    before_action :authenticate_user!
 
-    def authenticate_admin
-      # TODO: Add authentication logic here.
-    end
+    include ApplicationBaseModule
+
+    # def authenticate_admin
+    # TODO: Add authentication logic here.
+    # end
 
     # Override this value to specify the number of elements to display at a time
     # on index pages. Defaults to 20.
