+++ app/controllers/ingredients_controller.rb
@@ -1,10 +1,14 @@
+# frozen_string_literal: true
+
+# controller for Ingredients
 class IngredientsController < ApplicationController
+  before_action :authenticate_user!
+
   def show
-
-  end
-  def new
+    @ingredients = IngredientDecorator.decorate(Ingredient.all)
   end
 
-  def edit
-  end
+  def new; end
+
+  def edit; end
 end
