+++ app/helpers/ingredients_helper.rb
@@ -1,2 +1,4 @@
+# frozen_string_literal: true
+
 module IngredientsHelper
 end
