+++ spec/controllers/ingredients_controller_spec.rb
@@ -1,19 +1,19 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 
 RSpec.describe IngredientsController, type: :controller do
-
-  describe "GET #new" do
-    it "returns http success" do
+  describe 'GET #new' do
+    it 'returns http success' do
       get :new
       expect(response).to have_http_status(:success)
     end
   end
 
-  describe "GET #edit" do
-    it "returns http success" do
+  describe 'GET #edit' do
+    it 'returns http success' do
       get :edit
       expect(response).to have_http_status(:success)
     end
   end
-
 end
