+++ spec/decorators/ingredient_decorator_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 
 RSpec.describe IngredientDecorator do
