+++ spec/views/ingredients/edit.html.erb_spec.rb
@@ -1,5 +1,7 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 
-RSpec.describe "ingredients/edit.html.erb", type: :view do
+RSpec.describe 'ingredients/edit.html.erb', type: :view do
   pending "add some examples to (or delete) #{__FILE__}"
 end
